fix type issues

+++ onnx/reference/ops/op_lp_pool.py
@@ -38,4 +38,4 @@ class LpPool(CommonPool):
         )
 
         kernel_element_count = np.prod(kernel_shape)
-        return (np.power(kernel_element_count * power_average[0], 1.0 / p),)
+        return (np.power(kernel_element_count * power_average[0], 1.0 / p).astype(x.dtype),)
